requirements.txt
python-telegram-bot
python-dotenv
python-docx
fpdf2
markdown
streamlit
